Simplify logic to disable/enable checkbox for 2d reduction
Now that the GUI automatically switches reduction, unless the user
manually overrides it, we can simplify the logic.

This is because it's not obvious what the "correct" option is for 2D
reductions with CanSASNexus. Do we uncheck the tick-box? If so do we
re-check it with memory mode on when they switch to 1D or when they
switch to file mode.

As the default options (which switch automatically) are fine for 90% of
users lets leave the complexity out and let a power user make their own
decisions.

In the next commit we will improve the error message trying to process
2D data with CanSasNexus enabled