Make Simulation/DensityOfStates more robust if missing Euphonic
Previously, the missing Euphonic dependency would crash Mantid if a
.yaml or .castep_bin file was selected in a fresh
Indirect/Simulation/DensityOfStates window.

This is because the (nonexistent) "__dos_ions" would be accessed even
if an error was raised. I have changed the if/else logic to forbid
this, and the interface seems to perform as expected now.

- If Euphonic is uninstalled, there will be an error message related
  to failing to load ions, and in the warning text is information
  about the missing library.
- If it is installed, then .yaml and .castep_bin files will work.

I am not very familiar with the code for Mantid Interfaces so please
forgive/expect relevant oversights!