Make the PropertyManager optional
The PropertyManager is only needed to determine the wavelength range if
the user did not supply it. This changes the logic so it is no longer
needed. Also fixes a bug when a user supplies only min OR max wavelength
that it now only calculates the one that wasn't supplied.