Remove a couple of project save related errors
Neither error was blocking any functionality but removes distracting
messages:
- error about widget having already been deleted on shutdown of a unit test.
This was caused by the ADS being cleared before the UI was closed. I've added
a tearDown method so the UI is closed first
- error during save of project while the Engineering UI was open with Focus having
been run. This was due to the grouping workspace not being saved. We don't need it
to be saved since the project save is meant for the Fitting tab rather than Focus
tab so I've just added some better handling for grouping workspaces to avoid the red
Error message in the log