Add euphonic to Conda dependencies
Euphonic is available on Conda forge, which resolves potential
difficulties building it consistently with Numpy/Scipy. (This library
includes a compiled C extension and interfaces with Lapack.)

- Pinned to 0.6.* to enable force constants import via
  euphonic.cli.utils (this is likely to be moved in a
  future version).

For non-Conda builds we will try to get things working with Pip
wheels, but as Conda is the future we might as well get this in place
now.