Bump Sphinx to 1.8.5 on MSVC
Bumps MSVC to use Sphinx 1.8.5 (matching Ubuntu 18.04). This allows us
to build dev-docs with warnings enabled on both platforms.

RHEL is the odd one out now, but as we have (comparitvely) few
developers on RHEL and this only affects dev docs we can cross that
bridge when Conda is complete