Fix problem with sorting of events
The LoadEventNexus algorithm is run without compression and the events
in the Nexus file are not already sorted by pulse time, the algorithm
is incorrectly labelling them as sorted by pulse time. This change fixes
the problem by always looping through the "used" detectors even if
loading without compression
This problem was causing FilterByTime to incorrectly slice up an event
file by time. A lot of events were assigned to the last second of a run
when slicing by time