Bump cppcheck warning count up
Bumps cppcheck warning count up by 3. Cppcheck and clang-tidy disagree
on const ref and std::move.

Clang-tidy states that const ref makes the lifetime non-obvious, and
someone who want's to transfer ownership can use std::move

Cppcheck states that we should take a const ref because copy is
expensive (but doesn't not factor in std::move)

On balance, I'll favour clang-tidy because move is a thing and it
clarifies lifetime especially with shared_ptr. Since these warnings are
in a mess of 1700+ I can't see where the suppression goes, so lets leave
it for the future