Fix missing assignment op / copy ctor
Fixes instances with missing assignment operator or copy constructor.
These are dangerous cases with dynamic memory as it's not obvious what
will happen, but the compiler will try anyway.

In many cases we can simply switch to unique_ptr or disable the copy
when it makes sense. But one or two cases required additional thought.

This fixes several memory leaks or sources of corruption if the fields
fall out of sync