Fix silly != / == bug in beam center finder
Fixes a silly bug, where we checked for the front being not set, rather
than being set, for validity.
This meant when the value was set, we reverted back to using the rear
value.

Add a number of tests to cover this and various similar bugs we see in
on SANS just in-case.
Re #32541