Fix problem with applying pulse time filtering
The pulse time filtering previously assumed the events were sorted by
pulse time in the input Nexus file. This isn't always the case
The previous approach of assuming you could convert the pulse time
filter into an event id from\to range wasn't valid so I've just called
the FilterByTime algorithm as a post process. This probably isn't as
quick as the original method but don't think the original approach can be
fixed.
There's another filter (FilterDuringPause) set up as a template method
on the EventWorkspaceCollection class so I've done the same for the
FilterByTime method so it works on multi-period files. But since FilterByTime
doesn't have a feature to apply the filter in place I had to create a new
method on EventWorkspaceCollection
I noticed that LoadEventNexus currently passes m_ws->getSingleHeldWorkspace()
to filterDuringPause at one point instead of simply passing m_ws. I don't
think this is correct because it only results in the first workspace (period)
being filtered but am leaving it because not part of this bug fix