Plot result of loading workspace from ADS
Previously when a workspace existed in the ADS it was not hooked up to
actually continue to perform the plot. This commit refactors the code
slightly so that the presenter can call load from ADS first and plot the
result immediately if it succeeds; otherwise it starts loading async and
waits for a callback before doing the plotting, as per the existing
implementation.

Re #32490