Enable /bigobj on MSVC by default
This enables the bigobj flag for targets where the number of sections in
an obj file exceeds 2^16 for ISISReflectometryTestTargets. This is
because the underlying target pulls in Qt and MantidQtWidgetsCommon
pushing us just over the threshold.

This is enabled by Microsoft by default for UWP apps (so has first party
support). The drawback is linkers pre-2005 cannot read from these
objects, something that will not affect us:
https://stackoverflow.com/a/15110915