Fix SANS not producing a range list when parsing files
New TOML files allow us to accept and parse a range of log/lin
wavelength bins.
The old parser never accepted this, so a long-standing bug where the
GUI didn't update parser values from the user's input went unnoticed.

This commit will now construct a string from the list of wavelengths in
the GUI if we don't have the user's input.