Change instrument actor to run algorithms without the ADS
When using instrument actor from another GUI we are now using workspace
pointers rather than storing workspaces in the ADS so the instrument
actor needs to support this. As part of this we need to store a
non-const shared_ptr to the workspace rather than a weak_ptr. This makes
more sense anyway because the instrument actor modifies the workspace
e.g. when doing masking, it was just getting around this before by using
the ADS.

Re #32490