Hook up shape-changed notification and revert to edit mode
I will replace "pan" mode with "edit" mode so for now the name "pan" is
misleading, but this commit hooks up the callback from the shape-changed
signal in the view to notify the presenter that a shape has been
created. The presenter then reverts to "pan" mode (really edit mode) so
that the user can start editing the created shape. This ties in with
default instrument viewer behaviour. I will fix the misleading naming
and icon in a future commit.

Re #32566