Refactoring to rename "pan" to "edit"
The pan button has been replaced by a shape-editing button and this
commit renames the related items and functions to reflect this. It also
does some tidying to put the buttons (and related functions etc) in the
same order as they appear on the instrument viewer, i.e. zoom, edit,
select

Re #32566