Fix unreliable tests related to framework initialisation
Some expectations on errors had been added to the tests but these errors
were not genuine, they were actually due to the framework manager not
being initialised. This caused some unreliable tests where the framework
would sometimes be initialised and the errors were not produced.

This commit initialises the framework in the tests so we can remove the expectations on the errors.