Revert change to sort det id lists
ExtractSpectra resorts a workspace if the detector id list contains the
same detector ids as are present in the input workspace but they're sorted
differently. This is a useful feature if the V and sample workspace happen
to be sorted differently so it makes sense to allow this to happen.
Comparing the common_det_ids to the V and sample detector id lists without
sorting will pick up any sort order differences