Euphonic install systemtest: try using a newer Pip
The build is failing because Pip cannot build spglib on the Windows
CI; it claims it "cannot be installed directly" and the system does
not have the requirements for a local build. (Specifically, Python.h
is missing.)

Perhaps a newer version of Pip is able to "install directly" from
whatever packaging scheme spglib is using at the moment?