Fix segfault in instrument viewer pick tab
Prevent the miniplot in the pick tab from being updated if the tab is
not visible - ie when it is folded to the left

This solves the issue of the miniplot not being reachable by the update
functions, which caused the segfault. It has the unintended side effect
of drawing a shape with the tab already folded not updating the miniplot
when it is unfolded, but this is a bizarre use case, and I don't see how
to do any better right now.