Fix NoneType error is app_closing run twice
This is mainly to solve an issue on the build servers that I could not
reproduce locally on release/debug on any platform. But the fact it happened
on jenkins suggests it's possible so it's  worth adding the check just in
case.

Also removes some useless test file lines added in error.