Fix OSX build with OCCT 7.6.0
As of this commit https://github.com/Open-Cascade-SAS/OCCT/commit/a8b605eb5e7f181c7bed7c7ef89108c4fd7ba054 access to the internal
array of pointers underlying the triangle nodes has been removed. Instead we need to ask for the number of nodes (which we already
do) then simply call a getter for that index. This simplifies code on our side too.

There's a place where we are manually doing the normal of the triangles. Whilst the framework has a method built in, I'm not
comfortable changing this. Long-term we're going to have to migrate from OpenGL anyway on OSX so....