Make ReflectometryISISPreprocess work with groups
This algorithm was failing if the input is a workspace group because the output property was a MatrixWorkspaceProperty, which does not allow the output to be a group. This commit changes the output to be a WorkspaceProperty so that groups are supported.

As well as adding a test for this algorithm I've also added one for ReflectometryISISLoadAndProcess to ensure that the grouped output from ReflectometryISISPreprocess is dealt with correctly by that algorithm.