Change return type of onSurface
Previously it returned 0 for "no" and 1 for "yes." Since this is very
much what a bool is, just change to the language native object. The only
exception was Torus::onSurface which was hard coded to always return -1.
This would be interpreted as "true," but the comments in the code
suggest that it was meant to always return false.