Replace flakey InstrumentView view test with presenter tests
Replaces a view test (which should just assume the layer underneath
works, as we shouldn't have logic in views) which is currently failing
with the presenter tests.

Checking the view there isn't any major logic and the test is failing
due to us trying to start / interact with Qt widgets in a test thread.

These replacement tests were written using coverage to ensure I got 100%
of the presenter whilst maintaining bug compat.