Remove unused suppressions from cppcheck
Adding MantidKernel/SuppressWarnings apparently will cause cppcheck to
skip the downstreams cpp and .h files. This means that everything that
implicitly now has the warning suppression is no longer flagging
warnings at all.

As this has been a long-standing problem within the project, I suspect
we have a number of files with missed warnings because of this. For that
reason I think this commit should stand, and a future series should fix
the suppression then re-instate these warnings plus all the other ones
in the project