Abins: add multiplication methods to SData
This will allow energy-dependent corrections (which may actually be
q-dependent...) to be applied to calculated S intensities with clean
syntax.

The core logic of spowdersemiempiricalcalculator needs to juggle a few
different scenarios so it is helpful if the data manipulation can be
clear and terse - making it easier to focus on the logic.