Abins: apply appropriate q-dependence to autoconvolution spectra
High-order phonon contributions are estimated by repeated convolution
of the fundamental spectrum against the highest calculated order. (As
this may be the first order, we refer to the method as
'autoconvolution'. But strictly-speaking only the first step is
actually "auto"...)

In Abins energy and q are related by the instrument/angle geometry,
so the q^2 dependence of the spectrum is "baked in" to the calculation
over energies at each angle. However, as high order spectra should
scale with (q^2n) (where q is the quantum order), it is not correct
to rely on the q information in the fundamental spectrum. Instead,
the approach should be:

- Compute the spectra at a fixed q (1 recip. angstrom is
  convenient...)

- Apply the q^2n terms after q-independent line shapes are determined
  by convolution

Here we do this slightly inefficiently to fit the existing Abins
toolchain (further refactors to be considered...) and

- Compute the order-1 spectrum as a function of q (with and without
  the order-1 Debye-Waller factor) and, if desired, the order-2 spectrum.

- Remove the q-dependence from the non-DW-corrected spectra by
  division

- Use the fixed-q order-1 spectrum as the basis for auto-convolution

- Apply (q^2n / n!) intensity correction to each kernel

- Apply q-dependent Debye-Waller factors exp(-tr(A) q^2/3) to all
  orders

- Replace order-1 data with the previously-computed
  spectrum (including a mode-dependent Debye-Waller term.)

This replaces an implementation in which the n! correction was
neglected and convolution kernels were normalised to 1. It is believed
that some other implementations of this method perform convolution
with normalised kernels; however it seems more convenient to leave
the (tr(B)/3) term in the convolution than to remove it and later
apply (tr(B)/3)^n. Perhaps this comes at a cost in precision?