Experimental fix for invalid pulsetimes in sans data
This adjusts a previous workaround where pulsetimes before the epoch were adjusted. This assumed that the invalid pulsetimes were at the end of the data set and an estimate was based on that assumption. However we now have data where the invalid pulsetimes are not at the end. For now try simply excluding them. We also have pulsetimes that are after the epoch but before the run start time and these are excluded now too.