Make fit guess obtain settings from fit options browser
This commit makes the fit guess in muons fit using parameters in the fit options browser. This fixes an issue in the previous commit that caused a crash if the guess parameters had not been updated when the plot guess was clicked.
This pr also hides guess parameter rows when not required.

re: #31876