Resolve hidden bug with SANS system tests
The framework clean up was automatically clearing the ADS for us as part
of shutdown. Now we defer that to later it's highlighted a bug where the
SANS code tries to delete each workspace manually.
However this does not account for group workspaces removing more than 1
ws at a time (which was introduced with the wavelength changed), causing
the loop to fail when an expected ws is deleted implicitly