Exposed horizontal/vertical grouping to the auto-process algorithm
This commit renames GroupPixelsBy and connects properties for
horizontal and vertical grouping in DirectILLCollectData.
GroupDetBy is used solely to decrease granularity of data and groups pixels vertically,
inside of a tube. It worth to keep it separate from horizontal/vertical grouping intended
for increasing statistics for the flat background calculation, as in the latter case
the granularity reduction may be much larger than desired for the final data.