Fix type error toggling between ortho/non-ortho with lineplots
Previously in the toggle to orthogonal plots included a call to
"add_line_plot" if a line plot already existed. This was to preserve a
users lineplots if they had previously added them.

We have concluded this is an unused branch of code, as it would throw a
type error from missing params, yet nobody has noticed / reported it. As
the fix is non-trivial (it requires types only in the presenter), we
will remove it and users will have to re-toggle their line plots on in
the obscure case where they already have them on then completely change
projection type.