Replace internal Inelastic fortran modules
Use external quasielasticbayes package instead.
This uses the same sources but allows the package
to be installed externally on other platforms where
prebuilt binaries are not shipped with mantid

Some bugs in the fortran code were fixed in the
external package. This has given small changes in
peak heights but not shapes.