Add test interface to open RegionSelector
This is a hack so we can allow interested users / devs to test the
interface without us having to embed it at this point.

Simply add:
mantidqt.python_interfaces = Utility/RegionSelectorInterface.py
To your Mantid.user.properties to see the interface under the utility
tab. You'll need at least one ws in the ADS for this to work.