Abins2D: Add ILL-PANTHER instrument
Resolution function based on fitting to data from ILL.

It looks rather broad, need to double check that's ok...
But the main issue here is if we use an unphysically large energy
range. Constraining incident energy to the known instrument range
should help manage this. The next step would be to overhaul energy/bin
handling in general; currently Abins constructs energy bins all over
the place by referring to limits in abins.parameters. This is not
flexible enough to handle these cases without a lot of redundancy;
instead we could consider having the Algorithm specify a set of bins
which are then passed around.

This might be a good time to also drop "bin_width" in favour of n_bins?