Notify presenter when view is closed
When notified, the presenter clears the reference to the view. This
means we can simplify some of the logic in refresh_view that was
previously using sip to check if the view still exists.

We can also now test that the presenter does the correct thing on close.
This replaces the previous test that was in the view so the view test
file has now been deleted.