Implement custom move and copy constuctors
Due to the nature of the pointer from row to group, if the group was
moved the memory address that row held on to would be invalid. These
changes create custom move and copy constructors such that whenever the
group is moved or copied, the group gives each of its rows its new
memory address.

RE #26531