Improve manual testing instructions for the refl gui
Fix some minor confusing language.
Change the Save ASCII instructions to be specific for each output file format.
Chagne the Exit Prompt section to match current behaviour - it was previously documented that there is a Save/Discard/Cancel option but this is not the case; there is only OK/Cancel.