Couple of changes
Switch from using GroupDetectors to DiffractionFocussing when focussing distributions.
The fact that GroupDetectors logic used a common set of bins across all groups was
creating a dog-leg in the focussed data patterns for PEARL
Also modify the code in DiscusMultipleScatteringCorrection which converts an input
S(Q) profile from bin edges to points. The original method was causing an offset in
the corrected S(Q) profiles where the peaks where moved to the right v slightly