[RDEXP] Expose useThread to python
Because using a thread causes problems when the user clicks too fast, we
want not to use any. But exposing this argument to python requires it to
come before the dependencies arg (or at least it is infinitely easier to
convince sip to work this way round), so we refactor a bit the
instrument view constructor and its uses.

Also removes a closeEvent function defined twice for some reason.