Ignore errors when removing recovery checkpoints
Various error reports note an issue with removing
non-empty recovery checkpoint directories where
a WinError 145 is raised, e.g.

File "C:\MantidNightlyInstall\bin\workbench\app\mainwindow.py", line 603, in closeEvent
   self.project_recovery.remove_current_pid_folder()
 File "C:\MantidNightlyInstall\bin\workbench\projectrecovery\projectrecovery.py", line 199, in remove_current_pid_folder
   self._remove_directory_and_directory_trees(self.recovery_directory_pid)
 File "C:\MantidNightlyInstall\bin\workbench\projectrecovery\projectrecovery.py", line 138, in _remove_directory_and_directory_trees
   shutil.rmtree(directory)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 730, in rmtree
   return _rmtree_unsafe(path, onerror)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 603, in _rmtree_unsafe
   _rmtree_unsafe(fullname, onerror)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 612, in _rmtree_unsafe
   onerror(os.rmdir, path, sys.exc_info())
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 610, in _rmtree_unsafe
   os.rmdir(path)

This is hard to reproduce but is likely a timing issue when
stopping project recovery as the removal of the checkpoints
does not check the the latest recovery operation has
actually stopped. The recovery is then writing new files
while then trying to remove the directory.

We cannot do much with this scenario as workbench is closing
so errors are swallowed in the knowledge that future runs
of workbench will cleanup any checkpoints older than
1 month so at worse a small amount of disk space is used
for a while.