Separate abort of recovery thread and checkpoint removal
Several user-submitted error reports contain errors
when trying to remove the checkpoint directories
saying that they are not empty:

 File "C:\MantidNightlyInstall\bin\workbench\app\mainwindow.py", line 588, in closeEvent
   self.project_recovery.remove_current_pid_folder()
 File "C:\MantidNightlyInstall\bin\workbench\projectrecovery\projectrecovery.py", line 199, in remove_current_pid_folder
   self._remove_directory_and_directory_trees(self.recovery_directory_pid)
 File "C:\MantidNightlyInstall\bin\workbench\projectrecovery\projectrecovery.py", line 138, in _remove_directory_and_directory_trees
   shutil.rmtree(directory)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 730, in rmtree
   return _rmtree_unsafe(path, onerror)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 603, in _rmtree_unsafe
   _rmtree_unsafe(fullname, onerror)
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 612, in _rmtree_unsafe
   onerror(os.rmdir, path, sys.exc_info())
 File "C:\MantidNightlyInstall\bin\lib\shutil.py", line 610, in _rmtree_unsafe
   os.rmdir(path)

The likely cause is the recovery thread still running and
creating new files while the main thread has moved on to
try and delete the checkpoints. This is a low-risk fix
targeted at 6.3 to provide a little extra time for the
thread to finish before attempting the folder deletion.

A future will attempt to instead schedule the delete once
the recovery thread has finished.