Remove last modified dates from algorithm docs pages
These dates were parsed from the git log of the source
files and would include *any* change to the file regardless
of whether the change was meaningful, e.g. a formatting
change.
The preferred option going forward is for each page to
document any change in behaviour across versions in its
own changelog on the page.