Fix long_mode
Activating long_mode in a call to create_vanadium or focus wasn't having
any effect eg the long mode TOF cropping values weren't kicking in
I have fixed this by simplifying _apply_temporary_inst_settings so the
settings for the appropriate long mode are always applied regardless
of whether the long_mode has changed or not. This aligns the logic with the
PDF\Rietveld Mode in the POLARIS scripts
I have also changed the logic that applies atten=False when long mode is True
so that it uses the update_attributes function. This will generate a warning
for the user that the perform_attenuation setting has been changed as well as
updating the value