Abins: fix logic for interpolation between two kernels
This was not experienced when testing 1D examples where there was
always a decent range of broadening widths. However, when using a
Pychop instrument at low E_init it is possible that only two sigma
samples are needed for interpolation. In that case the existing logic
is broken, e.g.::

  n_kernels = ceil( log(1.3) / log(1.41) ) = 1

To interpolate properly we will need two kernels, one at min(sigma)
and one at min(sigma) * 1.41. But the n=1 case is caught by
``if n_kernels == 1:`` and we get a single kernel width [min(sigma)].

This is the correct outcome for a _constant_ width, in which case::

  n_kernels = ceil( log(min_sigma / min_sigma) / log(1.41) )
            = ceil(0.) = 0

To fix and clarify the logic, move the (+ 1) into n_kernels before the
if statement.