Abins: refactor algorithm inheritance, cut Abins2D redundancy
The main issue being solved here is the maintenance problem of needing
to assign local properties in any class derived from AbinsAlgorithm.

The opportunity is also taken to move "Autoconvolution" into the
common parameter set as the feature is available in both Abins and Abins2D.

There is an extensive use of class variables to store state, which
seems problematic; it would lead to strange behaviour if we try to run
multiple instances of Abins simultaneously.

I haven't eliminated all of those here, but taken the opportunity to
move things into __init__ statements while cutting the redundant elements.