Abins2D: Convert output units
ConvertUnits isn't working for this, so we just adjust the bin values
when creating workspaced. It doesn't really matter that everything was
calculated in cm-1...

Well, actually there is a problem which is that the spectrum units are
not properly normalised. Spectral values should really be something
like "barn/energy" and so vary with energy units. But Mantid likes to
normalise spectra to counts/bin. As we haven't changed the size of the
bin or the incoming z-values, the spectrum value also stays the same.

The problem is that this makes the values sensitive to the bin size
used in calculation; reasonable in an experimental context but it
doesn't deliver the correct underlying property. Abins2D spectra already
have different intensity with different bin size... Abins(1D) doesn't
have that problem. Why is it different? I don't see an explicit correction.