Various changes to support iterative MS calculation
Performance improvements in importance sampling functions:
- Skip prepareCumulativeProbForQ if not doing importance sampling
- Use vectors instead of histogram
- Use Histogram constructor without e in another place
- Use auto& instead of auto in a couple of places
- Reuse the same vectors in the bin loop to save allocating\freeing memory

Improve validation on S(Q) by checking for nans

Add CompAssembly to instrument helper function:
This means the instrument can be saved to a Nexus using
SaveNexusESS. SaveNexusESS should probably save the instrument
without requiring this but it seems like it requires the detectors
to be grouped into banks

Make algorithm more tolerant when given bad data - Check that wavelength
values are positive numbers

Modify the code which converts an input S(Q) profile from bin edges to points.
The original method was causing an offset in the corrected S(Q) profiles where
the peaks where moved to the right v slightly. Note - this caused one of the
unit tests to fail:
Need the spike to coincide with one of the detector two theta's for test to
work nicely and this has unfortunately moved it away

So move the spike in the test S(Q) to a new Q value. Note the most likely Q is still
slightly to the right of the S(Q) peak centre because the probabilty goes like
Q.S(Q) but close enough to the peak centre for the arithmetic to be valid
Also improved the windowing in the peak finding in the unit test

Make GetXMinMax work off bin centres rather than bin edges

Expose the start x and bin width in create2DWorkspaceWithGeographicalDetectors