Refactor LookupTable into a class
Change LookupTable from a vector of rows to its own class. This is under
the hood work to merge it with LookupRowFinder. For now the
implementation is in the LookupRow header because this is interim work
that will be removed, but this gets the class in with minimal changes to
ensure everything else still works.

Re #33564