Fix clang warning on mac
Refactoring from a previous PR meant that a check on the theta angle was
left in when it should have been removed - previously this was an
optional but now it is the actual value. The refactoring meant that a
valid value is always passed to this function now so the conditional is
no longer required.

Refs #33539