Abins2D: Converty Pychop resolution from FWHM to sigma
Abins/Pychop resolutions appear too broad compared to experiment, by a
factor of about 2. Presumably this is because we are
interpreting a full-width-half-maximum value as the Gaussian sigma;
these are related by a scale factor close to 2.35.

We should check how the resolution for PANTHER was defined; this is
likely to also be a FWHM but I haven't got this written explicitly.