Add hard-coded tooltips for lookup criteria columns
The lookup table columns that specify inputs to the reduction algorithm
take their tooltips from the algorithm property of the same name.
Previously the theta-lookup column also took the tooltip from the
algorithm. However, this is not the actual property that will be passed
to the reduction; instead it is lookup criteria for the other settings
that will be used.

This commit changes the tooltip to a hard-coded one for the lookup
column to make this clear. This will also make it more flexible for
adding additional lookup criteria that do not "accidentally" match a
similar looking algorithm property.

The column property names have been changed to a map and moved to the
view, which is the only place they are used.

Re #33563