Handle 11-col and future batch files in Refl
This squashes multiple commits:

Add version to encoder / decoder

Each time we've changed the Batch JSON we potentially need to support
older files. Guessing this from row count...etc. is brittle so lets add
version support whilst we are here to make this easier for future devs.

Re #33563

Fix the 11 column case to be aware of new version

Updates the various tests to include the version number, and implicitly
fixes old and new tests together.

Re #33563