MonteCarloAbsorption: Allow getting bounding box from environment
Previously, in case the beam shape has not been specified, a test
was performed whether the sample shape was valid. This may fail in
case only the container shape is defined as sample environment but
not the sample. In case the environment exist, but sample shape is
not valid, the bounding box is now extracted from the environment.