[RDEXP] Hack to keep the focus on the rdexp
Since whenever a new widget is open, or whenever a matplotlib plot si
changed, the widget steals the focus and render the keyboard interaction
impossible, the focus is immediately given back to the rdexp.

This hack is used because the event generated by this focusing out is
the same as clicking, so no discrimination here, and matplotlib cannot
be convinced not to steal the focus in the first place.