Fix and guard against various alg memory leaks
This follows the C++ Core Guidelines to help identify leaking
destructors:

- Mark all algorithms as final
Most algorithms are standalone, and this completely removes the need for
developers and compilers to consider inheritance

- Mark destructor for all inherited algs as protected
This allows us to check if the virtual destructor is required, by
preventing callers with the base pointer calling the destructor
implicitly (this will result in a compile time error). See cppcore
guidelines:
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-dtor-virtual

- Mark remaining destructors as virtual
By using the above technique, we can spot derived types being deleted
via their base pointer without a virtual d'tor. These will leak as we
will release base memory and not the dervied instance