Fixup file finding on Windows
On Windows the FileFinder does not consider a directory the same as a
file, unlike POSIX systems. Therefore we need to provide the folder and
filename so a regular file can be found.

This resolves the issues seen on Windows. Additionally, I've marked the
files as std::string types to make debugging this unit test easier in
the future.

Re #33580