Handle multiple row matches in exp. settings
This adds handling for if we match multiple theta AND title values.
Since we don't want to write our own checking logic in the view, we do
this once a user hits process so that we can check the regex values.

This is typically seen if a user has a greedy regex (e.g. ".*") and a
more specific matcher than conflict.

Re #33516