Abins: update CRYSTAL MgO-GX test displacement data
It turns out that the Gamma-point displacements do need updating to
match the other updated test file (mgo-GX_LoadCRYSTAL_data.txt). This
is due to an existing change in Abins; zero/negative-frequency modes
are no longer omitted from the loaded AbinsData, and instead are
filtered out during processing. This makes data shapes more regular
and possible loss of scientific data to a somewhat arbitrary
thresholding process.

To avoid completely regenerating test data (and hiding unintended
changes that might have occurred at the same time) the test was
written to be somewhat robust against old reference files, screening
out low frequencies with the _cull_imaginary_modes
staticmethod. However, this method was written such that the reference
frequencies and displacements were still expected to be /consistent/.

As in this case we have regenerated the *_data.txt file, to capture a
genuine change in how k-points are handled, the consistency was broken
at the Gamma-point, where there are (as expected) three imaginary
modes. So we need to update the displacements file at this k-point to
match.

I have checked the numbers by hand, and the relationship between the
files is as expected: new rows are added to the start of the file
which represent rigid lattice translations, and the existing values
are unchanged. (The linebreaks are also lost, but this should be
inconsequential.)