Fix a bug where autoreduction would end early
If a group only contains rows that fail due to the new error that is
caught, then an empty algorithms list would be returned. This would
cause autoreduction to assume that everything had been processed even
though there could be remaining groups to process. This commit ensures
we skip over groups that do not contain any valid rows.

Re #33516