Add warning when using wildcard row
With the added complexity of having two lookup rows, it could be
beneficial to warn the user when the run does not match any of the rows
in the experiment tab and so has hit the wildcard row. This should
reduce the effects of typos in the title field.

RE #33515