Fix bug when loading batch over existing lookup rows
If the lookup table already contained a number of lookup rows the
loading of a batch resulted in some of the rows that had been there
before sticking around due to an indexing error. This simply sets the
number of rows to 0 before loading, which gracefully handles the
clearing of the table under the hood.

RE #33585