Attempt image detach several times before failing
It has been observed that attempting to detach the
created image too quickly can lead to "resource busy"
errors and the build failing.
It is likely that other processes on the system, e.g.
Finder, Antivrius are keeping the image busy for a while.
We attempt to detach 5 times and then quit.