Fix a crash with reflectometry live data and sliceviewer
A crash was happening related to the fact that the ReflectometryReductionOneLiveData algorithm updates the input workspace. If sliceviewer is open on the workspace while this happens it is possible that it will cause mantid to crash. While the sliceviewer issue still needs resolving, the algorithm does not need to (and therefore should not) modify the input workspace anyway. This commit therefore removes this modification, which fixes the crash.

Re #33610