Abins: cache instrument setting data on getter
Instrument setting is immutable, so this only affects the case where
advanced parameters are changed after the Instrument was
instantiated - which Abins never does.

As well as saving some overhead, this removes the excessive logging
when using default setting. (Currently a message is logged every time
a parameter is checked with these methods...)